requirements.txt
Django==3.2
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
PyJWT==2.4.0
mysqlclient==2.1.1
drf-yasg==1.21.5
django-utils-six